refactor: 유틸리티 함수 및 컴포넌트에서 경로 수정 및 새로운 유틸리티 파일 추가
